Update code with some minor renaming to make imports easier
