+++ setup.py
@@ -2,7 +2,7 @@
 from setuptools import setup
 
 setup(
-    name='emulator-1ts',
+    name='emulator_1ts',
     version='0.0.1',
     python_requires='>3.10',
     packages=['emulator_1ts'],
